+++ python/ray/serve/tests/test_cli.py
@@ -10,11 +10,10 @@ import httpx
 import pytest
 import yaml
 
-import ray
 from ray import serve
 from ray._common.test_utils import wait_for_condition
 from ray.serve._private.common import DeploymentID
-from ray.serve._private.constants import SERVE_DEFAULT_APP_NAME, SERVE_NAMESPACE
+from ray.serve._private.constants import SERVE_DEFAULT_APP_NAME
 from ray.serve.scripts import remove_ansi_escape_sequences
 from ray.util.state import list_actors
 
@@ -30,46 +29,6 @@ def assert_deployments_live(ids: List[DeploymentID]):
         assert any(prefix in actor_name for actor_name in running_actor_names), msg
 
 
-def test_start_shutdown(ray_start_stop):
-    subprocess.check_output(["serve", "start"])
-    # deploy a simple app
-    import_path = "ray.serve.tests.test_config_files.arg_builders.build_echo_app"
-
-    deploy_response = subprocess.check_output(["serve", "deploy", import_path])
-    assert b"Sent deploy request successfully." in deploy_response
-
-    wait_for_condition(
-        check_http_response,
-        expected_text="DEFAULT",
-        timeout=15,
-    )
-
-    ret = subprocess.check_output(["serve", "shutdown", "-y"])
-    assert b"Sent shutdown request; applications will be deleted asynchronously" in ret
-
-    def check_no_apps():
-        status = subprocess.check_output(["serve", "status"])
-        return b"applications: {}" in status
-
-    wait_for_condition(check_no_apps, timeout=15)
-
-    # Test shutdown when no Serve instance is running
-    ret = subprocess.check_output(["serve", "shutdown", "-y"], stderr=subprocess.STDOUT)
-    assert b"No Serve instance found running" in ret
-
-
-def test_start_shutdown_without_serve_running(ray_start_stop):
-    # Test shutdown when no Serve instance is running
-    ret = subprocess.check_output(["serve", "shutdown", "-y"], stderr=subprocess.STDOUT)
-    assert b"No Serve instance found running" in ret
-
-
-def test_start_shutdown_without_ray_running():
-    # Test shutdown when Ray is not running
-    ret = subprocess.check_output(["serve", "shutdown", "-y"], stderr=subprocess.STDOUT)
-    assert b"Unable to shutdown Serve on the cluster" in ret
-
-
 def check_http_response(expected_text: str, json: Optional[Dict] = None):
     resp = httpx.post("http://localhost:8000/", json=json)
     assert resp.text == expected_text
@@ -77,9 +36,9 @@ def check_http_response(expected_text: str, json: Optional[Dict] = None):
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_deploy_basic(ray_start_stop):
+def test_deploy_basic(serve_instance):
     """Deploys some valid config files and checks that the deployments work."""
-    ray.init(address="auto", namespace=SERVE_NAMESPACE)
+    # ray.init(address="auto", namespace=SERVE_NAMESPACE)
 
     # Create absolute file names to YAML config files
     pizza_file_name = os.path.join(
@@ -152,43 +111,13 @@ def test_deploy_basic(ray_start_stop):
         assert_deployments_live(deployments)
         print("All deployments are live.\n")
 
-    ray.shutdown()
+    # ray.shutdown()
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_deploy_with_http_options(ray_start_stop):
-    """Deploys config with host and port options specified"""
-
-    f1 = os.path.join(
-        os.path.dirname(__file__), "test_config_files", "basic_graph_http.yaml"
-    )
-    success_message_fragment = b"Sent deploy request successfully."
-
-    with open(f1, "r") as config_file:
-        config = yaml.safe_load(config_file)
-
-    deploy_response = subprocess.check_output(["serve", "deploy", f1])
-    assert success_message_fragment in deploy_response
-
-    wait_for_condition(
-        lambda: httpx.post("http://localhost:8005/", json=None).text
-        == "wonderful world",
-        timeout=15,
-    )
-
-    # Config should contain matching host and port options
-    info_response = subprocess.check_output(["serve", "config"])
-    info = yaml.safe_load(info_response)
-
-    # TODO(zcin): the assertion should just be `info == config` here but the output
-    # formatting removes a lot of info.
-    assert info == config["applications"][0]
-
-
-@pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_deploy_multi_app_basic(ray_start_stop):
+def test_deploy_multi_app_basic(serve_instance):
     """Deploys some valid config files and checks that the deployments work."""
-    ray.init(address="auto", namespace=SERVE_NAMESPACE)
+    # ray.init(address="auto", namespace=SERVE_NAMESPACE)
 
     # Create absolute file names to YAML config files
     two_pizzas = os.path.join(
@@ -278,11 +207,11 @@ def test_deploy_multi_app_basic(ray_start_stop):
         assert_deployments_live(deployment_names)
         print("All deployments are live.\n")
 
-    ray.shutdown()
+    # ray.shutdown()
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_deploy_duplicate_apps(ray_start_stop):
+def test_deploy_duplicate_apps(serve_instance):
     """If a config with duplicate app names is deployed, `serve deploy` should fail.
     The response should clearly indicate a validation error.
     """
@@ -299,7 +228,7 @@ def test_deploy_duplicate_apps(ray_start_stop):
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_deploy_duplicate_routes(ray_start_stop):
+def test_deploy_duplicate_routes(serve_instance):
     """If a config with duplicate routes is deployed, the PUT request should fail.
     The response should clearly indicate a validation error.
     """
@@ -316,7 +245,7 @@ def test_deploy_duplicate_routes(ray_start_stop):
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_deploy_bad_v2_config(ray_start_stop):
+def test_deploy_bad_v2_config(serve_instance):
     """Deploy a bad config with field applications, should try to parse as v2 config."""
 
     config_file = os.path.join(
@@ -336,7 +265,7 @@ def test_deploy_bad_v2_config(ray_start_stop):
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_deploy_multi_app_builder_with_args(ray_start_stop):
+def test_deploy_multi_app_builder_with_args(serve_instance):
     """Deploys a config file containing multiple applications that take arguments."""
     # Create absolute file names to YAML config file.
     apps_with_args = os.path.join(
@@ -367,7 +296,7 @@ def test_deploy_multi_app_builder_with_args(ray_start_stop):
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_config_multi_app(ray_start_stop):
+def test_config_multi_app(serve_instance):
     """Deploys multi-app config and checks output of `serve config`."""
 
     # Check that `serve config` works even if no Serve app is running
@@ -390,7 +319,7 @@ def test_config_multi_app(ray_start_stop):
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_cli_without_config_deploy(ray_start_stop):
+def test_cli_without_config_deploy(serve_instance):
     """Deploys application with serve.run instead of a config, and check that cli
     still works as expected.
     """
@@ -414,12 +343,12 @@ def test_cli_without_config_deploy(ray_start_stop):
         return True
 
     wait_for_condition(check_cli)
-    serve.shutdown()
-    ray.shutdown()
+    # serve.shutdown()
+    # ray.shutdown()
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_config_with_deleting_app(ray_start_stop):
+def test_config_with_deleting_app(serve_instance):
     """Test that even if one or more apps is deleting, serve config still works"""
 
     config_json1 = {
@@ -481,7 +410,7 @@ def test_config_with_deleting_app(ray_start_stop):
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_status_basic(ray_start_stop):
+def test_status_basic(serve_instance):
     """Deploys a config file and checks its status."""
 
     # Check that `serve status` works even if no Serve app is running
@@ -537,7 +466,7 @@ def test_status_basic(ray_start_stop):
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_status_error_msg_format(ray_start_stop):
+def test_status_error_msg_format(serve_instance):
     """Deploys a faulty config file and checks its status."""
 
     config_file_name = os.path.join(
@@ -564,7 +493,7 @@ def test_status_error_msg_format(ray_start_stop):
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_status_invalid_runtime_env(ray_start_stop):
+def test_status_invalid_runtime_env(serve_instance):
     """Deploys a config file with invalid runtime env and checks status.
 
     get_status() should not throw error (meaning REST API returned 200 status code) and
@@ -589,7 +518,7 @@ def test_status_invalid_runtime_env(ray_start_stop):
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_status_syntax_error(ray_start_stop):
+def test_status_syntax_error(serve_instance):
     """Deploys Serve app with syntax error, checks error message has traceback."""
 
     config_file_name = os.path.join(
@@ -612,7 +541,7 @@ def test_status_syntax_error(ray_start_stop):
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_status_constructor_error(ray_start_stop):
+def test_status_constructor_error(serve_instance):
     """Deploys Serve deployment that errors out in constructor, checks that the
     traceback is surfaced.
     """
@@ -640,7 +569,7 @@ def test_status_constructor_error(ray_start_stop):
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_status_constructor_retry_error(ray_start_stop):
+def test_status_constructor_retry_error(serve_instance):
     """Deploys Serve deployment that errors out in constructor, checks that the
     retry message is surfaced.
     """
@@ -668,7 +597,7 @@ def test_status_constructor_retry_error(ray_start_stop):
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_status_package_unavailable_in_controller(ray_start_stop):
+def test_status_package_unavailable_in_controller(serve_instance):
     """Test that exceptions raised from packages that are installed on deployment actors
     but not on controller is serialized and surfaced properly.
     """
@@ -692,7 +621,7 @@ def test_status_package_unavailable_in_controller(ray_start_stop):
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_max_replicas_per_node(ray_start_stop):
+def test_max_replicas_per_node(serve_instance):
     """Test that max_replicas_per_node can be set via config file."""
 
     config_file_name = os.path.join(
@@ -716,7 +645,7 @@ def test_max_replicas_per_node(ray_start_stop):
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_replica_placement_group_options(ray_start_stop):
+def test_replica_placement_group_options(serve_instance):
     """Test that placement group options can be set via config file."""
 
     config_file_name = os.path.join(
@@ -741,7 +670,7 @@ def test_replica_placement_group_options(ray_start_stop):
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-def test_deploy_from_import_path(ray_start_stop):
+def test_deploy_from_import_path(serve_instance):
     """Test that `deploy` works from an import path."""
 
     import_path = "ray.serve.tests.test_config_files.arg_builders.build_echo_app"
@@ -762,32 +691,74 @@ def test_deploy_from_import_path(ray_start_stop):
 
 
 @pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
-@pytest.mark.parametrize(
-    "ray_start_stop_in_specific_directory",
-    [
-        os.path.join(os.path.dirname(__file__), "test_config_files"),
-    ],
-    indirect=True,
-)
-def test_deploy_with_access_to_current_directory(ray_start_stop_in_specific_directory):
-    """Test serve deploy using modules in the current directory succeeds.
+def test_status_multi_app(serve_instance):
+    """Deploys a multi-app config file and checks their status."""
+    # Check that `serve status` works even if no Serve app is running
+    subprocess.check_output(["serve", "status"])
+    print("Confirmed `serve status` works when nothing has been deployed.")
 
-    There was an issue where dashboard client doesn't add the current directory to
-    the sys.path and failed to deploy a Serve app defined in the directory. This
-    test ensures that files in the current directory can be accessed and deployed.
+    # Deploy config
+    config_file_name = os.path.join(
+        os.path.dirname(__file__), "test_config_files", "pizza_world.yaml"
+    )
+    subprocess.check_output(["serve", "deploy", config_file_name])
+    print("Deployed config successfully.")
 
-    See: https://github.com/ray-project/ray/issues/43889
-    """
-    # Deploy Serve application with a config in the current directory.
-    subprocess.check_output(["serve", "deploy", "use_current_working_directory.yaml"])
-
-    # Ensure serve deploy eventually succeeds.
-    def check_deploy_successfully():
+    def num_live_deployments():
         status_response = subprocess.check_output(["serve", "status"])
-        assert b"RUNNING" in status_response
-        return True
+        status = yaml.safe_load(status_response)["applications"]
+        return len(status["app1"]["deployments"]) and len(status["app2"]["deployments"])
 
-    wait_for_condition(check_deploy_successfully, timeout=5)
+    wait_for_condition(lambda: num_live_deployments() == 3, timeout=15)
+    print("All deployments are live.")
+
+    status_response = subprocess.check_output(
+        ["serve", "status", "-a", "http://localhost:8265/"]
+    )
+    statuses = yaml.safe_load(status_response)["applications"]
+
+    expected_deployments_1 = {"f", "BasicDriver"}
+    expected_deployments_2 = {
+        "Multiplier",
+        "Adder",
+        "Router",
+    }
+    for deployment_name, deployment in statuses["app1"]["deployments"].items():
+        expected_deployments_1.remove(deployment_name)
+        assert deployment["status"] in {"HEALTHY", "UPDATING"}
+        assert "message" in deployment
+    for deployment_name, deployment in statuses["app2"]["deployments"].items():
+        expected_deployments_2.remove(deployment_name)
+        assert deployment["status"] in {"HEALTHY", "UPDATING"}
+        assert "message" in deployment
+    assert len(expected_deployments_1) == 0
+    assert len(expected_deployments_2) == 0
+    print("All expected deployments are present in the status output.")
+
+    for status in statuses.values():
+        assert status["status"] in {"DEPLOYING", "RUNNING"}
+        assert time.time() > status["last_deployed_time_s"]
+    print("Verified status and deployment timestamp of both apps.")
+
+
+@pytest.mark.skipif(sys.platform == "win32", reason="File path incorrect on Windows.")
+def test_deployment_contains_utils(serve_instance):
+    """Test when deployment contains utils module, it can be deployed successfully.
+
+    When the deployment contains utils module, running serve deploy should successfully
+    deployment the application and return the correct response.
+    """
+
+    config_file = os.path.join(
+        os.path.dirname(__file__),
+        "test_config_files",
+        "deployment_uses_utils_module.yaml",
+    )
+
+    subprocess.check_output(["serve", "deploy", config_file], stderr=subprocess.STDOUT)
+    wait_for_condition(
+        lambda: httpx.post("http://localhost:8000/").text == "hello_from_utils"
+    )
 
 
 if __name__ == "__main__":
